Add ability to provide file name mapper to to_ndjson_dir
